🔥 chore: rename getRunScriptCertbotCommands to getCertbotCommands
